Overhaul of Globals.h and a_inits.ino
Made change in Globals.h to allow for 20T or 16T gear, ring versions and moved all the values over to a_inits.ino to keep the clutter down in the globals file.  Added error check for incorrect values entered in for gear size and ring version.  Mismatched gears could be added but why would anyone use that?  Maybe it's just my OCD but I see no reason to add them to the code.